comments and separate entrypoint functions
