+++ linera-explorer/README.md
@@ -1,6 +1,12 @@
-## Linera Explorer
+# Linera Explorer
 
-### Build instructions
+<!-- cargo-rdme start -->
+
+This module provides web files to run a block explorer from linera service node.
+
+<!-- cargo-rdme end -->
+
+## Build instructions
 
 After building linera with `cargo`, run
 ```bash
@@ -11,3 +17,11 @@ and then
 webpack
 ```
 It will create a `dist` folder from where you can run a HTTP server
+
+## Contributing
+
+See the [CONTRIBUTING](../CONTRIBUTING.md) file for how to help out.
+
+## License
+
+This project is available under the terms of the [Apache 2.0 license](../LICENSE).
